[Chore/#6] Code Review 반영사항 반영
