+++ Daangn_iOS/Daangn_iOS/Network/DTO/ProductResponseDOT.swift
@@ -1,0 +1,19 @@
+//
+//  ProductResponseDOT.swift
+//  Daangn_iOS
+//
+//  Created by 정정욱 on 11/19/24.
+//
+
+import Foundation
+
+struct Product {
+    let thumbnailImageName: String
+    let title: String
+    let distance: String
+    let location: String
+    let time: String
+    let price: String
+    let chatCount: Int
+    let likeCount: Int
+}
